ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
pydantic==2.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.26.0
# shap==0.44.1  # Optional: SHAP explanations (requires cmake/llvmlite, install manually if needed)
imbalanced-learn==0.12.0
scipy==1.12.0
pyarrow==15.0.0
